Move history/snapshot stuff into history folder
