+++ app/src/lib/history/SnapshotCard.svelte
